+++ sources/vector.c
@@ -1,21 +1,21 @@
 #include "doom_nukem.h"
 
-void		set_vector(t_v3d *v, float x, float y, float z)
+void		set_vector(t_v3d *v, double x, double y, double z)
 {
 	v->x = x;
 	v->y = y;
 	v->z = z;
 }
 
-float 		vector_length(t_v3d v)
+double 		vector_length(t_v3d v)
 {
-	return (sqrtf(vector_dot_product(v, v)));
+	return (sqrt(vector_dot_product(v, v)));
 }
 
 t_v3d	vector_normalise(t_v3d v)
 {
 	t_v3d	result;
-	float		length;
+	double		length;
 
 	length = vector_length(v);
 	result.x = v.x / length;
@@ -24,7 +24,7 @@ t_v3d	vector_normalise(t_v3d v)
 	return (result);
 }
 
-float		vector_dot_product(t_v3d v1, t_v3d v2)
+double		vector_dot_product(t_v3d v1, t_v3d v2)
 {
 	return (v1.x * v2.x +
 			v1.y * v2.y +
